Make en_short consistent with other sites
This applies three changes to en_short:

1. Removes spaces between numbers and units, so that "1 h" becomes "1h"
2. Shortens "min" to "m", so that "1 min" becomes "1m" (month remains as "mo")
2. Shortens "yr" to "y", so that "1 yr" becomes "1y"

I made these changes because this is what I've witnessed to be the standard on the rest of internet. This format is used by Facebook, Twitter, LinkedIn, Venmo, and others.

Feel free to reject some or all changes if this doesn't match with the direction you want, I'll be happy to revert individual parts.